chatbot/requirements.txt
streamlit
google-generativeaiS